Add ability to define auth_page
It's a page where user will be redirected if he or she
don't special cookie. It's used to give user a chooice
if he/she would like to use basic auth or OAuth variant.

Can be useful to improve support of 'satisfy any;' in Nginx.

E.x. some of employes don't have github account but must have access
to the system.